Added test file and made the split file as a header file which will reduce code length in single file
