+++ split_command_and_flags.h
@@ -1,11 +1,3 @@
-#include <iostream>
-#include <string>
-#include <cctype>
-#include <vector>
-#include <algorithm>
-
-using namespace std;
-
 vector <string> split(const string &text, char sep) {
 	vector<string> tokens;
 	size_t start = 0, end = 0;
@@ -18,13 +10,3 @@ vector <string> split(const string &text, char sep) {
 	return tokens;
 }
 
-int main() {
-
-	// Get the string using " getline(cin, test); "
-	string test = "  ls   -l  ";
-	vector < string > command = split(test, ' ');
-	for (int i = 0;  i < command.size(); ++i) {
-		cout << command[i] << endl;	
-	}
-	return 0;
-}
